requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
gunicorn==19.6.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==3.2
